allow the PG cache to handle cache return values changing (#340)
The API currently needs to be restarted whenever we update the model
hosts enum. This bothered me enough to see if it was fixable, and it
turns out there's a connection option in postgres that fixes this! This
query param will allow our PG driver to roll back and skip the cached
statement.

https://jdbc.postgresql.org/documentation/use/#connection-parameters

> autosave (String) Default never
Specifies what the driver should do if a query fails. In autosave=always
mode, JDBC driver sets a savepoint before each query, and rolls back to
that savepoint in case of failure. In autosave=never mode (default), no
savepoint dance is made ever. In autosave=conservative mode, savepoint
is set for each query, however the rollback is done only for rare cases
like ‘cached statement cannot change return type’ or ‘statement XXX is
not valid’ so JDBC driver rolls back and retries