Fix message delete when label exists. (#372)
I noticed in production the delete endpoint throwing an error related to
sqlalchemy

```
(psycopg.errors.NotNullViolation) null value in column "message" of relation "label" violates not-null constraint
DETAIL:  Failing row contains (lbl_G5M6F5K9H5, null, 1, google-oauth2|115800053695087261993, null, 2025-09-04 18:13:20.175565+00, null).
[SQL: UPDATE label SET message=%(message)s::VARCHAR WHERE label.id = %(label_id)s::VARCHAR]
[parameters: {'message': None, 'label_id': 'lbl_G5M6F5K9H5'}]
(Background on this error at: https://sqlalche.me/e/20/gkpj)
```

This is happening because cascade was not set on messages table and
there is a foreign key on label. With cascade set, sqlalchemy tries to
update the label to a null value and fails causing the delete endpoint
to not work for messages with labels.

This adds a fix and test for this issue.