Safety check settings (#387)
Adds a flag that allows the user to bypass the google safety checks, if
they have the permission to do so.

There is a new permission that is added called
`write:bypass-safety-check` . Open to changing this name.

Closes: https://github.com/allenai/playground-issues-repo/issues/606

TODO: see if i can get a test user that doesn't have the correct
permission for testing. Currently only have tests for anonymous and
users with the permission.