Add OpenTelemetry, including Flask and Pydantic-AI instrumentation (#389)
Closes https://github.com/allenai/playground-issues-repo/issues/304

Minimal setup for Otel tracing the backend. I want to build on this, but
also want to get a small working version out to verify it works in
production. Locally it seems to be working well and makes it really easy
to see what is going on with the api calls. It will be very easy to move
over different timing related metrics we have in the app.



With docker compose running you can go to 

http://localhost:16686/search

and look for spans. Things seem to be mostly working, expect that the
pydantic span is not getting added to the api span and i can't figure
out why.

<img width="2515" height="255" alt="Screenshot 2025-09-22 at 3 42 24 PM"
src="https://github.com/user-attachments/assets/91fcf84e-7c00-40ac-878b-aef9209141a7"
/>



Some guides i have been looking at 
https://ai.pydantic.dev/logfire/#otel-without-logfire
https://ai.pydantic.dev/direct/#when-to-use-the-direct-api-vs-agent


https://opentelemetry-python.readthedocs.io/en/stable/examples/fork-process-model/README.html