add per model default inference options (#388)
The PR add inference constraints (default, min, max, and step) to the
model_config table and related endpoints as a flat list of fields (see
below). They are initialized in the table by sql with the typical
defaults from the old v3/schema. e2e tests added for create model and
update model.

```
        temperatureDefault,
        temperatureLower,
        temperatureUpper,
        temperatureStep,
        topPDefault,
        topPLower,
        topPUpper,
        topPStep,
        maxTokensDefault,
        maxTokensLower,
        maxTokensUpper,
        maxTokensStep,
        stopDefault,
  ```

 issues: 
 https://github.com/allenai/playground-issues-repo/issues/577
 https://github.com/allenai/playground-issues-repo/issues/627

---------

Co-authored-by: Taylor Blanton <matthewtblanton@gmail.com>