Adds ErrorChunk to stream and persists error message when encountered (#405)
This PR adds an ErrorChunk chunk to the stream which is a way to return
a message to the UI that contains properties for alerting inline in the
message thread based on the errorSeverity sent. When the error chunk is
detected during streaming the stream is exited early and the error is
persisted to the db as a Message. Open Telemetry span errors are
recorded where the ErrorChunk is created.

This currently only addresses one error case: 
* when the LLM sends a tool call for a tool that doesn't exist.

Future work:
* refactor so that error chunk doesn't exit the stream early and stop
pending assistant responses

This goes with these UI change:
https://github.com/allenai/olmo-ui/pull/1214