Pydantic ai video patch (#418)
This PR creates a subclass of OpenAIChatModel called
OpenAIChatModelVideo which supports video in inputs.

To support video we have to change a large method inside of the
OpenAIChatModel called _map_user_prompt . This method isn't meant to be
changed and there is no guaranteed that it would change in the future.
Updating Pydantic ai could break this code.

Closes: https://github.com/allenai/playground-issues-repo/issues/707