Video safety check (#415)
This PR implements the Google video safety checker. 

It also
* Adds a private storage bucket used to pass data to the safety checker.
* Removes  safety_check_elapsed_time in favor of using Otel
* is_msg_harmful, which was always true or none (the conditions that
made it false cause an error to be thrown)


This PR causes the video to be uploaded twice. Once to the safety bucket
and once approved to our molmo bucket. There might be a small
optimization here to instead move it from one bucket to another, but
doing this requires some extra complexity. I believe in production this
cost will be pretty small, as its an upload within the the same zone. I
also didn't want to do this yet incase we end up changing our video
safety solution.

The other major limitation of the google video api is that it only
detects explicit content. We might want something that can categorize
more things.

closes https://github.com/allenai/playground-issues-repo/issues/699