Add index to label.message (#410)
While looking into issues with the migration we noticed that a labels
query was taking a while. Since it searches for `label.message` we
thought that column could use an index.

This PR makes that index and also sets up a naming convention in
SQLAlchemy for indexes. We can do it for other things like fkeys but
that seems like it would mess up older fkeys we've made.