make infini-gram sources optional, refactor to use pydantic output (#426)
Closes https://github.com/allenai/playground-issues-repo/issues/797

This resolves issues we were having when calling attribution for Olmo 3
indexes. We were trying to get data about the source that didn't exist
in our configs. This sets a default of `None` for the things that would
get mapped that don't exist.

This also updates the attribution endpoint to use the OpenAPI mapping
correctly so the UI can use it.