Handle molmo 2 points as part of input_parts (#433)
Addresses https://github.com/allenai/playground-issues-repo/issues/804

The API part of https://github.com/allenai/olmo-ui/pull/1261

This handles the proposed `inputParts` input for molmo 2 point input
parts.

If we want, we can save these parts to the DB and format them when we
send the message to the model. Would make it much easier to have
different UI for them!