Remap input parts to correct type. (#435)
This is a short term fix to correct the data in input_parts. 

In line 51 in threads_blueprint, we dump the request into dictionaries
and loose all the classes.

For most things this doesn't matter because we map this into a pydantic
class in line 113 of endpoint.

However input parts are used before this final mapping so the type is
incorrect.

We should probably fix the type of ModelMessageStreamInput or make it
into a pydantic class ,but i'll leave that for another day.