check mimetypes instead of extensions for followups (#436)
This PR addresses some issues we're having with followups. 

Instead of checking for a set list of file extensions, this uses
Python's `mimetype` library to get the MIME type. It uses the MIME type
to determine what kind of Pydantic part to map to.

This is more flexible and will update whenever Python does.