Save input parts to the DB (#434)
It seemed like it would be nice to keep the input parts around instead
of turning them into text immediately. This PR makes that happen.

It adds a Pydantic mapper to go to/from the DB so we don't have to work
too hard there.

The mapping to text content from the input parts now happens in the
pydantic mapping layer, not before.