Limit the safety worker process count (#454)
This seems to help with issues running the workers? Reducing the process
count will reduce the total # of redis connections and possibly reduce
memory use.

No clue why they're failing tbh, there's no logs about them shutting
down and no events. May have been a lot of memory use or something, but
I'm not sure why that wasn't happening in testing.

We can safely reduce process and thread count because most of the work
we're doing is async. Nothing will block for long.

This is inspired by this blog post:
https://medium.com/@haixinwang97/sunas-agent-running-system-a-practical-deep-dive-dramatiq-redis-on-kubernetes-43ae01bbefde